--- MailSender.lib/MailSenderService.cs
@@ -1,6 +1,0 @@
-﻿namespace MailSender.lib
-{
-    class MailSenderService
-    {
-    }
-}
