+++ MailSender/MailSender.csproj
@@ -91,6 +91,7 @@
     <Compile Include="Commands\ApplicationCloseCommand.cs" />
     <Compile Include="ValidationRules\RegExValidationRule.cs" />
     <Compile Include="ViewModel\MainWindowViewModel.cs" />
+    <Compile Include="ViewModel\SimpleIocExtentions.cs" />
     <Compile Include="ViewModel\ViewModelLocator.cs" />
     <Compile Include="View\RecipientsEditor.xaml.cs">
       <DependentUpon>RecipientsEditor.xaml</DependentUpon>
@@ -156,13 +157,6 @@
       <Generator>SettingsSingleFileGenerator</Generator>
       <LastGenOutput>Settings.Designer.cs</LastGenOutput>
     </None>
-  </ItemGroup>
-  <ItemGroup>
-    <Resource Include="Media\add.png" />
-    <Resource Include="Media\clock.png" />
-    <Resource Include="Media\edit.png" />
-    <Resource Include="Media\remove.png" />
-    <Resource Include="Media\scheduler.png" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\MailSender.lib\MailSender.lib.csproj">
