+++ MailSender/MainWindow.xaml
@@ -57,49 +57,10 @@
                     </Grid.RowDefinitions>
                     <ToolBarTray Grid.Row="0">
                         <ToolBar>
-                            <!--<Label MinWidth="85">Отправители</Label>
-                            --><!--<Separator></Separator>--><!--
-                            <ComboBox MinWidth="75" MaxWidth="150" ItemsSource="{x:Static data:TestData.Senders}" 
-                                      DisplayMemberPath="Name" SelectedIndex="1"/>
-                            <Button ToolTip="Создать">
-                                <Image Source="Media/add.png" Height="18"/>
-                            </Button>
-                            <Button ToolTip="Редактировать">
-                                <Image Source="Media/edit.png" Height="18"/>
-                            </Button>
-                            <Button ToolTip="Удалить">
-                                <Image Source="Media/remove.png" Height="18"/>
-                            </Button>-->
-                            <ctrl:ListController PanelName="Отправители" 
-                                                 Items="{x:Static data:TestData.Senders}"
-                                                 ViewPropertyPath="Name"/>
+                            <ctrl:ListController PanelName="Отправители" ViewPropertyPath="Name"/>
                         </ToolBar>
                         <ToolBar Band="1" >
-                            <!--<Label MinWidth="85">Сервера</Label>
-                            --><!--<Separator></Separator>--><!--
-                            <ComboBox MinWidth="75" MaxWidth="150" ItemsSource="{x:Static data:TestData.Servers}" SelectedIndex="0">
-                                <ComboBox.ItemTemplate>
-                                    <DataTemplate DataType="{x:Type entityes:Server}">
-                                        <StackPanel Orientation="Horizontal">
-                                            <TextBlock Text="{Binding Name}"/>
-                                            <TextBlock Text="{Binding Address}" Margin="3,0,0,0" Opacity="0.6"
-                                                       ToolTip="{Binding Port, StringFormat=Port:{0}}"/>
-                                            <TextBlock Text="{Binding Port, StringFormat=:{0}}" Opacity="0.6"/>
-                                        </StackPanel>
-                                    </DataTemplate>
-                                </ComboBox.ItemTemplate>
-                            </ComboBox>
-                            <Button ToolTip="Создать">
-                                <fa:ImageAwesome Height="18" Icon="Plus" Foreground="DarkBlue"/>
-                            </Button>
-                            <Button ToolTip="Редактировать">
-                                <fa:ImageAwesome Height="18" Icon="Edit" Foreground="DarkMagenta"/>
-                            </Button>
-                            <Button ToolTip="Удалить">
-                                <fa:ImageAwesome Height="18" Icon="Times" Foreground="Red"/>
-                            </Button>-->
-                            <ctrl:ListController PanelName="Сервера" 
-                                                 Items="{x:Static data:TestData.Servers}">
+                            <ctrl:ListController PanelName="Сервера">
                                 <ctrl:ListController.ItemTemplate>
                                     <DataTemplate DataType="{x:Type entityes:Server}">
                                         <StackPanel Orientation="Horizontal">
