+++ MailSender/View/RecipientsEditor.xaml
@@ -8,7 +8,6 @@
              DataContext="{Binding MainWindowModel, Source={StaticResource Locator}}"
              d:DesignHeight="365" d:DesignWidth="270">
     <Grid Margin="3">
-        <!-- Обработчик события ошибок валидации Validation.Error="OnDataValidationError" -->
         <Grid.Resources>
             <Style TargetType="{x:Type TextBox}">
                 <Setter Property="Margin" Value="5,2"/>
@@ -24,7 +23,6 @@
                                            Margin="3,0"
                                            ToolTipService.BetweenShowDelay="0"
                                            ToolTip="{Binding AdornedElement.(Validation.Errors)[0].ErrorContent, ElementName=AdornedElement}"/>
-
                                 <Border Padding="2" BorderBrush="Red" BorderThickness="3" CornerRadius="3">
                                     <AdornedElementPlaceholder Name="AdornedElement"/>
                                 </Border>
