+++ MailSender/View/RecipientsView.xaml
@@ -15,9 +15,6 @@
     </i:Interaction.Triggers>
     <DockPanel>
         <TextBlock DockPanel.Dock="Top" HorizontalAlignment="Center" Margin="0,10,0,7" FontSize="23" Text="Получатели писем">
-            <!--<i:Interaction.Behaviors>
-
-            </i:Interaction.Behaviors>-->
             <i:Interaction.Triggers>
                 <i:EventTrigger EventName="MouseUp">
                     <i:InvokeCommandAction>
@@ -42,17 +39,6 @@
                   AlternatingRowBackground="LightGray"
                   IsReadOnly="True" 
                   HeadersVisibility="Column">
-            <DataGrid.ContextMenu>
-                <ContextMenu>
-                    <MenuItem Header="Меню 1">
-                        <MenuItem Header="Подменю 1"/>
-                        <MenuItem Header="Подменю 2"/>
-                        <MenuItem Header="Подменю 3"/>
-                        <MenuItem Header="Подменю 4"/>
-                    </MenuItem>
-                    <MenuItem Header="Меню 1"/>
-                </ContextMenu>
-            </DataGrid.ContextMenu>
             <DataGrid.Columns>
                 <DataGridTextColumn Header="ID" Binding="{Binding Id}"/>
                 <DataGridTextColumn Header="Имя" Binding="{Binding Name}"/>
