+++ MailSender/ViewModel/SimpleIocExtentions.cs
@@ -1,0 +1,34 @@
+using System;
+using GalaSoft.MvvmLight.Ioc;
+
+namespace MailSender.ViewModel
+{
+    public static class SimpleIocExtentions
+    {
+        public static SimpleIoc TryRegister<TInterface, TService>(this SimpleIoc services)
+            where TInterface : class
+            where TService : class, TInterface
+        {
+            if (services.IsRegistered<TInterface>()) return services;
+            services.Register<TInterface, TService>();
+            return services;
+        }
+
+        public static SimpleIoc TryRegister<TService>(this SimpleIoc services)
+            where TService : class
+        {
+            if (services.IsRegistered<TService>()) return services;
+            services.Register<TService>();
+            return services;
+        }  
+
+        public static SimpleIoc TryRegister<TService>(this SimpleIoc services, Func<TService> Factory)
+            where TService : class
+        {
+            if (services.IsRegistered<TService>()) return services;
+            services.Register(Factory);
+            return services;
+        }
+
+    }
+}
