+++ MailSender/ViewModel/ViewModelLocator.cs
@@ -1,4 +1,3 @@
-using System;
 using CommonServiceLocator;
 using GalaSoft.MvvmLight.Ioc;
 using MailSender.lib.Data.Linq2SQL;
@@ -29,34 +28,5 @@ namespace MailSender.ViewModel
         }
 
         public MainWindowViewModel MainWindowModel => ServiceLocator.Current.GetInstance<MainWindowViewModel>();
-    }
-
-    public static class SimpleIocExtentions
-    {
-        public static SimpleIoc TryRegister<TInterface, TService>(this SimpleIoc services)
-            where TInterface : class
-            where TService : class, TInterface
-        {
-            if (services.IsRegistered<TInterface>()) return services;
-            services.Register<TInterface, TService>();
-            return services;
-        }
-
-        public static SimpleIoc TryRegister<TService>(this SimpleIoc services)
-           where TService : class
-        {
-            if (services.IsRegistered<TService>()) return services;
-            services.Register<TService>();
-            return services;
-        }  
-
-        public static SimpleIoc TryRegister<TService>(this SimpleIoc services, Func<TService> Factory)
-           where TService : class
-        {
-            if (services.IsRegistered<TService>()) return services;
-            services.Register(Factory);
-            return services;
-        }
-
     }
 }
