+++ Tests/MailSender.ConsoleTest/Program.cs
@@ -9,6 +9,4 @@ namespace MailSender.ConsoleTest
             Console.ReadLine();
         }
     }
-
-   
 }
